k8s: use NullableString instead of StringValue for preconditions
The UpdatePods API uses nullable string values to represent
annotations/labels preconditions. A null annotation value means that a
pod should only be updated if the corresponding annotation is not
currently set.

grpc-gateway deserializes a null `StringValue` as a empty string instead
of as a null pointer. This makes it impossible to use the current API to
express that an annotation should not be set.

This change introduces a new `NullableString` protobuf type to work
around this issue.