auditing: add local audit storage
This change pulls the audit storage into its own interface to decouple the SQL
implementation of audit storage from the basic audit service, and adds an
in-memory implementation of that interface. This allows for users to run Clutch
with some form of auditing, including sinks, without setting up a Postgres
server.

The new option is config backwards compatible. Documentation is added to the
in-depth security auditing docs, including clear language that this is not
recommended for production usage.