rtds: scope stats to xds version
This scopes the RTDS stats to the xDS version used in order to make it
easier to monitor the individual resources. Also adds a stat for errors
received from xDS clients.

An integration test is added to verify that we're emitting the
appropriate stats for V2 and V3. This test also serves to verify that
we're properly exposing the two protocols.