Add a new method to get aws.Config from DirectClient
AWS has too many services and it is impossible for clutch to support all
of them. If someone wants to write a private extension for a given
service, today there is no way to get the config/credentials to
construct a client. This patch exposes a new method to get regional
config so that clients to an unsupported service can be created.