frontend: Add full screen option to app configuration (#3196)
This pull request includes changes to the `AppLayout` component in the
`frontend/packages/core/src/AppLayout/index.tsx` file and the
`AppConfiguration` interface in the
`frontend/packages/core/src/Types/app.tsx` file. The primary goal of
these changes is to introduce the option for a full-screen layout.

Enhancements to layout configuration:

*
[`frontend/packages/core/src/AppLayout/index.tsx`](diffhunk://#diff-1465c1bdd75a440d3a5592c719280d01afe84948428c566711068196e9a02decL36-R45):
Modified the `AppLayout` component to conditionally render the header
and drawer based on the new `useFullScreenLayout` configuration option.
*
[`frontend/packages/core/src/Types/app.tsx`](diffhunk://#diff-cf066233a855942622a45960706a4b5e3bb99869f102c3e33c040d7983fa6862R10):
Added a new `useFullScreenLayout` boolean property to the
`AppConfiguration` interface to support the new layout option.