[backend] Update golangci-lint to support go 1.24 (#3265)
<!--- TITLE FORMAT: "component: short description", e.g. "k8s: add pod
log reader" -->

### Description
<!-- Describe your change below. -->
While trying to update the codebase to go 1.24 we noticed the
golangci-lint version was a bit outdated and causing some issues,
updating it and ran `make backend-lint` and `make backend-lint-fix` to
ensure the lint is correctly working again. While updating, also noticed
lots of `G115` exceptions in our backend / lint tests, so added a
`//nolint` comment to avoid having a linter failure while we check each
of the cases to see if it would cause an overflow.
<!-- Reference previous related pull requests below. -->

<!-- [OPTIONAL] Include screenshots below for frontend changes. -->

### Testing Performed
<!-- Describe how you tested this change below. -->

- Successfully running `make backend-lint`

### GitHub Issue
<!-- Link to any existing GitHub issues related to this PR. -->

Fixes #

### TODOs
<!-- Include any TODOs outstanding for the submission of this pull
request below. -->
- [ ] Check for overflows in the int conversions followed by a //nolint
comment
<!--
Example:
- [ ] This is an item on my TODO list.
- [x] This is a completed item.
-->